onecard-app/pyproject.toml
[project]
name = "onecard-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=45.0.6",
    "fastapi[standard]>=0.116.1",
    "psycopg2>=2.9.10",
    "python-jose[cryptography]>=3.5.0",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.41",
    "uvicorn[standard]>=0.35.0",
]
